Formating function declaration

+++ ft_libft/srcs/ft_isalnum.c
@@ -6,11 +6,11 @@
 /*   By: tsomacha <tsomacha@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 16:13:41 by tsomacha          #+#    #+#             */
-/*   Updated: 2024/11/04 16:53:48 by tsomacha         ###   ########.fr       */
+/*   Updated: 2025/03/28 06:23:36 by tsomacha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_isalnum(long int c)
+int	ft_isalnum(int c)
 {
 	if ((c >= 48 && c <= 57) || (c >= 65 && c <= 90) || (c >= 97 && c <= 122))
 	{
